backend/requirements.txt
Flask==2.3.3
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.7.0
marshmallow==3.20.1
python-dotenv==1.0.0
Flask==2.3.3
pymongo==4.6.0
bcrypt==4.1.0
PyJWT==2.8.0
marshmallow==3.20.1
boto3==1.34.55  # Para integração com AWS